requirements/stable.txt
#Core
Django

# WSGI HTTP Server
gunicorn
pytz

# Debugging
ipdb
ipython
django-debug-toolbar


# Configuration
django-environ
pillow

# Database
psycopg2
